core-components: split up LogViewer
